Tests/LibWeb/Text/expected/async-scrolling/smooth-scroll-in-nested-navigable.txt

immediate: 0
final: 300
